requirements.txt
notion-client==2.0.0
icalendar==5.0.7
Flask==2.3.3
python-dotenv==1.0.0
APScheduler==3.10.1
requests==2.31.0 